chore(types): finish the mypy ratchet — zero disabled error codes (#37) (#59)
Re-enable the last three deferred codes and fix all 54 errors:

- no-any-return (20): numpy/scipy scalars wrapped in float()/int()
  (float.__pow__ returns Any in typeshed; numpy 2.4 shape elements are
  Any); JSON loads and array returns cast(...) to precise targets;
  eval._band_mask return tightened to NDArray[np.bool_].
- type-arg (24): bare generics parameterized with the real types
  (spec: dict[str, Any] matching the FilterSpec.model_validate
  precedent; report_pdf._add_summary_page gets dict[str, list[Path]]).
- no-untyped-call (10): all genuinely third-party (schemdraw element
  constructors, matplotlib PdfPages.savefig) — targeted two-module
  [[tool.mypy.overrides]] with the libraries named, not per-line
  ignores.

The global disable_error_code list is now EMPTY — since #30 turned
mypy on (194 errors), correctness was enforced and completeness
deferred; both are enforced from here on.

Also: fix test_runs_headless (#58) to delete DISPLAY and render rather
than asserting the environment is headless — it failed on any dev
machine with X running.

Closes #37.

+++ packages/mcp-qucs-s/src/mcp_qucs_s/microstrip.py
@@ -63,7 +63,7 @@ def _eff_permittivity(w_h: float, er: float) -> float:
         + (1.0 / 18.7) * math.log(1.0 + (w_h / 18.1) ** 3)
     )
     b = 0.564 * ((er - 0.9) / (er + 3.0)) ** 0.053
-    return (er + 1) / 2 + (er - 1) / 2 * (1 + 10 / w_h) ** (-a * b)
+    return float((er + 1) / 2 + (er - 1) / 2 * (1 + 10 / w_h) ** (-a * b))
 
 
 def _impedance(w_h: float, er_eff: float) -> float:
